(PC-41934) refactor(location): remove useGeolocation
